test(runner): split failure recovery tests
