+++ crates/runner/src/cmd/start/tests/mod.rs
@@ -1,5 +1,5 @@
 mod budget;
-mod failure_panic;
+mod failure_recovery;
 mod idle_reuse;
 mod main_loop;
 mod support;
